move jwt to an external file
